Update instruction_execution.cpp
Added new output.
Final memory is stored in final_data.mc
Final registers are stored in final_state.mc
In each cycle if any register is updated, the value is stored in register_log.mc. Similarly, if any memory is updated, the value is stored in memory_log.mc. 
Also the format of simulator_log.txt is changed